Collaps refs in iter_asm if mutually deeper
I observed that given two references A and B, the 3' part of A and the 5' end
of B can get good depth. The low-depth regions then get sequencing errors.
This can lead to a situation where the two assemblies appear different, and
both have good mean depth and coverage, yet the two references should be
collapsed.

Now, parse KMA's mat.gz file to measure depth along segment length. For a ref
pair A, B, if there are big parts where A >> B and parts where B >> A, discard
one reference.